Print dapico-load output in formula test

+++ Formula/dapico-tools.rb
@@ -19,10 +19,13 @@ class DapicoTools < Formula
   end
 
   test do
+    return unless OS.mac?
+
     system "#{bin}/dapico-reboot", "--help"
     system "#{bin}/dapico-load", "--help"
     testpath.install resource("test-elf")
-    output = shell_output("#{bin}/dapico-load --dryrun #{testpath}/test.elf")
+    output = shell_output("#{bin}/dapico-load --dryrun #{testpath}/test.elf 2>&1")
+    puts "dapico-load output:\n#{output}"
     assert_match "write RAM 0x20000000 (33500 bytes", output
     assert_match "write RAM 0x200082dc (5832 bytes", output
     assert_match "execute at 0x200001e9", output
